requirements.txt
python-telegram-bot[job-queue]==22.7
openai==2.32.0
python-dotenv==1.2.2
python-docx==1.2.0
pymupdf==1.27.2.2
pillow==12.2.0
